extend attributes to 64 bits in API exposing them

+++ WinPort/WinCompat.h
@@ -466,7 +466,7 @@ typedef struct _SMALL_RECT {
 typedef struct _CONSOLE_SCREEN_BUFFER_INFO {
     COORD dwSize;
     COORD dwCursorPosition;
-    WORD  wAttributes;
+    DWORD64  wAttributes;
     SMALL_RECT srWindow;
     COORD dwMaximumWindowSize;
 } CONSOLE_SCREEN_BUFFER_INFO, *PCONSOLE_SCREEN_BUFFER_INFO;
